+++ modules/@apostrophecms/image/ui/apos/components/AposMediaManager.vue
@@ -105,7 +105,6 @@
           />
           <AposMediaManagerSelections
             :items="selected"
-            :can-edit="moduleOptions.canEdit"
             @clear="clearSelected" @edit="updateEditing"
             v-show="!editing"
           />
@@ -247,7 +246,8 @@ export default {
     async getMedia (options) {
       const qs = {
         ...this.filterValues,
-        page: this.currentPage
+        page: this.currentPage,
+        viewContext: this.relationshipField ? 'relationship' : 'manage'
       };
       const filtered = !!Object.keys(this.filterValues).length;
       if (this.moduleOptions && Array.isArray(this.moduleOptions.filters)) {
@@ -353,9 +353,7 @@ export default {
       this.editing = undefined;
     },
     async updateEditing(id) {
-      if (!this.moduleOptions.canEdit) {
-        return;
-      }
+      const item = this.items.find(item => item._id === id);
       // We only care about the current doc for this prompt,
       // we are not in danger of discarding a selection when
       // we switch images
@@ -370,7 +368,11 @@ export default {
           return false;
         }
       }
-      this.editing = this.items.find(item => item._id === id);
+      if (!item?._edit) {
+        this.editing = null;
+        return true;
+      }
+      this.editing = item;
       return true;
     },
     // select setters