+++ modules/@apostrophecms/doc/index.js
@@ -857,7 +857,7 @@ module.exports = {
       // Called by an `@apostrophecms/doc-type:insert` event handler to confirm that the user
       // has the appropriate permissions for the doc's type and content.
       testInsertPermissions(req, doc, options) {
-        if (!(options.permissions === false)) {
+        if (options.permissions !== false) {
           if (!self.apos.permission.can(req, 'create', doc)) {
             throw self.apos.error('forbidden');
           }