+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 ### Fixes
 
-* Pinned Vue dependency to 2.7.15. Released on December 24th, Vue 2.7.16 broke the rich text toolbar
-in Apostrophe.
+* Pinned Vue dependency to 2.7.15. Released on December 24th, Vue 2.7.16 broke the rich text toolbar in Apostrophe.
 
 ## 3.61.0 (2023-12-21)