+++ CHANGELOG.md
@@ -16,7 +16,9 @@
 
 * Fix the `launder` and `finalize` methods of the `float` schema field query builder.
 * Fix the `launder` and `finalize` methods of the `integer` schema field query builder.
-
+* A user who has permission to `publish` a particular page should always be allowed to insert it into the
+published version of the site even if they could not otherwise insert a child of the published
+parent.
 
 ## 3.61.1 (2023-01-08)