+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED
+## 3.62.0 (2024-01-25)
 
 ### Adds